enhance(placement): introduce a placment context
Previously the logic for fetching lived within the component.
By refactoring this we start to enable more complex queries for future
enhancements.
