+++ client/src/ui/organisms/placement/index.tsx
@@ -6,41 +6,17 @@ import { useUserData } from "../../../user-context";
 
 import "./index.scss";
 import { useGleanClick } from "../../../telemetry/glean-context";
+import {
+  PlacementData,
+  PlacementStatus,
+  usePlacement,
+} from "../../../placement-context";
 
 interface Timer {
   timeout: number | null;
   start: number | null;
   notVisible?: boolean;
 }
-
-enum Status {
-  success = "success",
-  geoUnsupported = "geo_unsupported",
-  capReached = "cap_reached",
-}
-
-export interface Fallback {
-  click: string;
-  view: string;
-  copy: string;
-  image: string;
-  by: string;
-}
-
-interface PlacementError {
-  status: Status.geoUnsupported | Status.capReached;
-}
-
-interface PlacementStatus {
-  status: Status.success;
-  click: string;
-  view: string;
-  copy?: string;
-  image?: string;
-  fallback?: Fallback;
-}
-
-type PlacementData = PlacementStatus | PlacementError;
 
 function viewed(
   pong: PlacementStatus,
@@ -57,45 +33,9 @@ function viewed(
 }
 
 export function Placement() {
-  const user = useUserData();
-  const gleanClick = useGleanClick();
-  const {
-    data: pong,
-    isLoading,
-    isValidating,
-  } = useSWR<PlacementData>(
-    !PLACEMENT_ENABLED || user?.settings?.noAds ? null : "/pong/get",
-    async (url) => {
-      const response = await fetch(url, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ keywords: [] }),
-      });
+  const pong = usePlacement();
 
-      gleanClick(`pong: pong->fetched ${response.status}`);
-
-      if (!response.ok) {
-        throw Error(response.statusText);
-      }
-
-      try {
-        const placementResponse: PlacementData = await response.json();
-        gleanClick(`pong: pong->status ${placementResponse.status}`);
-        return placementResponse;
-      } catch (e) {
-        throw Error(response.statusText);
-      }
-    },
-    {
-      revalidateIfStale: true,
-      revalidateOnFocus: false,
-      revalidateOnReconnect: false,
-    }
-  );
-
-  return isLoading || isValidating ? (
+  return !pong ? (
     <section className="place"></section>
   ) : (
     <PlacementInner pong={pong}></PlacementInner>
